% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-deepar.R
\name{deepar_fit_impl}
\alias{deepar_fit_impl}
\title{Bridge GluonTS DeepAR Modeling Function}
\usage{
deepar_fit_impl(
  x,
  y,
  freq,
  horizon,
  context_length = NULL,
  num_layers = 2,
  num_cells = 40,
  cell_type = "lstm",
  dropout_rate = 0.1,
  use_feat_dynamic_real = FALSE,
  use_feat_static_cat = FALSE,
  use_feat_static_real = FALSE,
  cardinality = NULL,
  embedding_dimension = NULL,
  distr_output = "default",
  scaling = TRUE,
  lags_seq = NULL,
  time_features = NULL,
  num_parallel_samples = 100,
  imputation_method = NULL,
  ctx = NULL,
  epochs = 100,
  batch_size = 32,
  num_batches_per_epoch = 50,
  learning_rate = 0.001,
  learning_rate_decay_factor = 0.5,
  patience = 10,
  minimum_learning_rate = 5e-05,
  clip_gradient = 10,
  weight_decay = 10 - 8,
  init = "xavier",
  hybridize = TRUE
)
}
\description{
Bridge GluonTS DeepAR Modeling Function
}
