% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-deepar.R
\name{deepar_predict_impl}
\alias{deepar_predict_impl}
\title{Bridge prediction Function for DeepAR Models}
\usage{
deepar_predict_impl(object, new_data)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}
}
\description{
Bridge prediction Function for DeepAR Models
}
