% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-deepstate.R
\name{deepstate_predict_impl}
\alias{deepstate_predict_impl}
\title{Bridge prediction Function for GP Forecaster Models}
\usage{
deepstate_predict_impl(object, new_data)
}
\arguments{
\item{object}{An object of class \code{model_fit}.}

\item{new_data}{A rectangular data object, such as a data frame.}
}
\description{
Bridge prediction Function for GP Forecaster Models
}
\keyword{internal}
