% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gluonts-env.R
\name{gluonts-env}
\alias{gluonts-env}
\alias{is_gluonts_activated}
\alias{activate_gluonts}
\alias{get_python_env}
\alias{check_gluonts_dependencies}
\alias{detect_default_gluonts_env}
\title{GluonTS Environment Connection Utilities}
\usage{
is_gluonts_activated()

activate_gluonts()

get_python_env()

check_gluonts_dependencies()

detect_default_gluonts_env()
}
\description{
These functions are used for testing and establishing a python environment
connection with \code{modeltime.gluonts}.
}
\details{
\itemize{
\item \code{is_gluonts_activated()}: Determines if a GluonTS Environment has been activated
during \code{library(modeltime.gluonts)}.
\itemize{
\item If \code{TRUE}, then you should be good to go.
\item If \code{FALSE}, then a connection between \code{modeltime.gluonts}
and your GluonTS Python has \emph{not} been activated.
}
\item \code{activate_gluonts()}: Attempts to activate a connection between \code{modeltime.gluonts}
and an associated GluonTS Python Environment using \code{reticulate::use_condaenv(required = TRUE)}.
\itemize{
\item It first looks for the system environment variable, 'GLUONTS_PYTHON', for a path to the python executable
\item It next looks for a Conda Environment named 'r-gluonts' (this is what most users will have)
}
\item \code{get_python_env()}: Returns the configuration for the python environment that is being discovered
using \code{reticulate::py_discover_config()}.
\item \code{check_gluonts_dependencies()}: Checks whether GluonTS required python dependencies are present in the
currently activated Python Environment.
\item \code{detect_default_gluonts_env()}: Detects if an 'r-gluonts' python environment is available.
\itemize{
\item Returns a \code{tibble} containing the
\item Returns \code{NULL} if an 'r-gluonts' environment is not detected
}
}
}
\examples{
\donttest{
# Returns TRUE if GluonTS connection established on package load
is_gluonts_activated()

#

}



}
\seealso{
\itemize{
\item \code{\link[=install_gluonts]{install_gluonts()}} - Used to install the python environment needed to run \code{modeltime.gluonts}.
}
}
