% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-deepar.R
\name{gluonts_deepar_fit_impl}
\alias{gluonts_deepar_fit_impl}
\title{Bridge GluonTS DeepAR Modeling Function}
\usage{
gluonts_deepar_fit_impl(x, y, freq, horizon, epochs)
}
\description{
Bridge GluonTS DeepAR Modeling Function
}
