% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_gluonts}
\alias{install_gluonts}
\title{Install GluonTS}
\usage{
install_gluonts()
}
\description{
Installs \code{GluonTS} Probabilisitic Time Series Modeling software
using \code{reticulate::py_install()}.
\itemize{
\item A \code{Python} Environment will be created
named \code{r-gluonts}.
\item The Modletime GluonTS R package will connect to the Python environment to use \code{GluonTS}
}
}
