% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-nbeats.R
\name{nbeats_ensemble_predict_impl}
\alias{nbeats_ensemble_predict_impl}
\title{Bridge prediction Function for N-BEATS ENSEMBLE Models}
\usage{
nbeats_ensemble_predict_impl(object, new_data)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}
}
\description{
Bridge prediction Function for N-BEATS ENSEMBLE Models
}
