% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-save_and_load.R
\name{save_gluonts_model}
\alias{save_gluonts_model}
\alias{load_gluonts_model}
\title{Saving and Loading GluonTS Models}
\usage{
save_gluonts_model(object, path, overwrite = FALSE)

load_gluonts_model(path)
}
\arguments{
\item{object}{A fitted model object}

\item{path}{A directory to store the GluonTS model files}

\item{overwrite}{Whether or not to allow overwriting a GluonTS model's directory. Default: FALSE.}
}
\description{
Saving and Loading GluonTS Models
}
