% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gluon_list_dataset.R
\name{to_gluon_list_dataset}
\alias{to_gluon_list_dataset}
\title{Convert a data frame to a GluonTS ListDataset}
\usage{
to_gluon_list_dataset(data, date_var, value_var, group_var = NULL, freq = "D")
}
\arguments{
\item{data}{A data.frame}

\item{date_var}{The date column (Timestamps)}

\item{value_var}{The value column (Target)}

\item{group_var}{The grouping column if Time Series Groups}

\item{freq}{the Pandas Timestamp Frequency.}
}
\description{
A ListDataset is the format required by \code{GluonTS}. This function
simplifies creating
}
\examples{
library(timetk)

m4_daily \%>\%
    to_gluon_list_dataset(
        date_var  = date,
        value_var = value,
        group_var = id,
        freq      = "D"
    )


}
