% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-trainer_params.R
\name{trainer_params}
\alias{trainer_params}
\alias{lookback_length}
\alias{num_batches_per_epoch}
\alias{learn_rate_decay_factor}
\alias{scale_values}
\title{General Tuning Parameters for GluonTS Trainers (All Models)}
\usage{
lookback_length(range = c(dials::unknown(), dials::unknown()), trans = NULL)

num_batches_per_epoch(range = c(2, 100), trans = NULL)

learn_rate_decay_factor(range = c(0, 1), trans = NULL)

scale_values(values = c(TRUE, FALSE))
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}

\item{values}{A character string of possible values.}
}
\description{
These parameters are part of the \code{gluonts.mx.trainer.Trainer()} function.
}
\details{
The main parameters for GluonTS Model Training are:
\itemize{
\item \code{epochs}: The number of iterations of training. See \code{dials::epochs()}.
\item \code{num_batches_per_epoch}: Number of batches at each epoch.
\item \code{learn_rate}: The rate at which learning is changed. See \code{dials::learn_rate()}.
\item \code{learn_rate_decay_factor}: Factor by which to decrease the learning rate.
\item \code{penalty}: The weight decay (or L2 regularization) coefficient. See \code{dials::penalty()}.
}

Additional useful parameters that have been added to Modeltime are:
\itemize{
\item \code{lookback_length}: Number of time units that condition the predictions.
\item \code{scale_values}: Scales numeric data by `id' group using mean = 0, standard deviation = 1 transformation.
}
}
