% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-install.R
\name{install_gluonts}
\alias{install_gluonts}
\title{Install GluonTS}
\usage{
install_gluonts()
}
\description{
Installs \code{GluonTS} Probabilisitic Deep Learning Time Series Forecasting Software
using \code{reticulate::py_install()}.
\itemize{
\item A \code{Python} Environment will be created
named \code{r-gluonts}.
\item The Modletime GluonTS R package will connect to the \code{r-gluonts} Python environment to use \code{GluonTS}
}
}
\examples{
\dontrun{
install_gluonts()
}


}
