% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-nbeats_params.R
\name{nbeats_params}
\alias{nbeats_params}
\alias{num_stacks}
\title{Tuning Parameters for NBEATS Models}
\usage{
num_stacks(range = c(2, 100), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}

\item{values}{A character string of possible values.}
}
\description{
Tuning Parameters for NBEATS Models
}
\details{
The main parameters for NBEATS models are:
\itemize{
\item \code{num_stacks}: The number of stacks the network should contain.
}
}
